# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aura Vault — REST API
  version: "1.0.0"
  description: |
    Backend REST API for the Aura Vault Protocol.

    ## Authentication
    All protected endpoints require a Bearer JWT access token obtained from
    `POST /api/auth/login`.  Access tokens expire after **15 minutes**.
    Use `POST /api/auth/refresh` to obtain a new pair without re-authenticating.

    ## Rate Limiting
    Auth endpoints (`/api/auth/*`) are limited to **20 requests per 15 minutes**
    per IP address.  Exceeding this returns `429 Too Many Requests`.

    ## Base URL
    | Environment | Base URL |
    |---|---|
    | Local | `http://localhost:3001` |
    | Testnet | `https://api-testnet.auravault.xyz` |
    | Mainnet | `https://api.auravault.xyz` |
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://api-testnet.auravault.xyz
    description: Stellar Testnet
  - url: https://api.auravault.xyz
    description: Stellar Mainnet
...
paths:
...
  /api/v1/vault/withdraw:
    post:
      tags: [Vault]
      summary: Submit a signed withdraw transaction
      description: |
        Submits a signed Soroban withdraw transaction envelope to Horizon.
        Supports `Idempotency-Key` header (UUID) for safe network retries.
      operationId: submitWithdraw
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
          description: Optional UUID idempotency key for safe retries (cached for 24 hours in Redis).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [signedXdr, shares, address]
              properties:
                signedXdr:
                  type: string
                shares:
                  type: string
                  example: "1000"
                address:
                  type: string
      responses:
        "200":
          description: Withdrawal submitted successfully
          headers:
            X-Idempotent-Replay:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  operation: { type: string, example: "withdraw" }
                  address: { type: string }
                  shares: { type: string }
                  hash: { type: string }
                  ledger: { type: integer }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Horizon transaction failed
...
  /api/v1/withdraw:
    post:
      tags: [Withdrawal]
      summary: Initiate withdrawal
      description: |
        For small withdrawals (≤ 100,000 shares) returns unsigned Soroban
        transaction parameters for the client to sign.  Large withdrawals are
        queued and return a `jobId` for status polling.
      operationId: initiateWithdrawal
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
            example:
              walletAddress: "GABCDE..."
              shares: "500"
              contractId: "CAURA_VAULT_TESTNET"
      responses:
        "200":
          description: Immediate withdrawal — unsigned tx params returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalImmediateResponse"
        "202":
          description: Large withdrawal queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalQueuedResponse"
              example:
                queued: true
                jobId: "550e8400-e29b-41d4-a716-446655440000"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/withdraw/queue/stats:
    get:
      tags: [Withdrawal]
      summary: Withdrawal queue stats
      description: Returns current queue depth and worker status for withdrawal processing.
      operationId: getWithdrawalQueueStats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/withdraw/{jobId}:
    get:
      tags: [Withdrawal]
      summary: Get withdrawal job status
      description: Poll the status of a queued large withdrawal.
      operationId: getWithdrawalJob
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Queue
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `POST /api/auth/login`.
        Pass as `Authorization: Bearer <token>`.
...
  schemas:
...
    ApiMeta:
      type: object
      description: Metadata attached to every API response envelope.
      required: [requestId, timestamp, version]
      properties:
        requestId:
          type: string
          format: uuid
          description: Unique identifier for the request, useful for log correlation.
          example: "550e8400-e29b-41d4-a716-446655440000"
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp when the response was generated.
          example: "2026-08-29T13:14:38.843Z"
        version:
          type: string
          description: API response schema version.
          example: "1.0.0"
...
    EnvelopeError:
      type: object
      description: |
        Standard error response envelope.
        Every error response is wrapped in this structure.
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          enum: [false]
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Machine-readable error code. One of the following values:

                | Code | Meaning |
                |---|---|
                | `INVALID_INPUT` | A required parameter was missing or had an invalid value |
                | `INTERNAL_ERROR` | An unexpected server-side error occurred |
                | `INVALID_ADDRESS` | The supplied address failed format validation |
                | `VAULT_PAUSED` | The vault is paused; no mutating operations are allowed |
                | `NOT_INITIALIZED` | The vault has not been initialised yet |
                | `ALREADY_INITIALIZED` | The vault has already been initialised |
                | `ZERO_AMOUNT` | A zero or negative amount was supplied |
                | `MATH_OVERFLOW` | An arithmetic operation overflowed |
                | `INSUFFICIENT_SHARES` | The caller does not hold enough vault shares |
                | `INSUFFICIENT_UNDERLYING` | The vault cannot cover the redemption |
                | `BALANCE_MISMATCH` | Actual on-chain balance differs from tracked state (flash-loan guard) |
                | `TVL_CAP_EXCEEDED` | The deposit would push the vault above its TVL cap |
                | `INVALID_WITHDRAWAL_FEE` | The withdrawal fee is outside the accepted range |
                | `FEE_EXCEEDS_MAXIMUM` | The requested fee exceeds the protocol maximum |
                | `HARVEST_COOLDOWN` | Harvest attempted before the cooldown period has elapsed |
                | `CIRCUIT_BREAKER_TRIPPED` | The circuit breaker has tripped; operation temporarily disabled |
                | `ASSET_NOT_FOUND` | The requested asset was not found |
                | `WEIGHT_MISMATCH` | The supplied asset weights do not sum to 100 % |
                | `DUPLICATE_ASSET` | The same asset appears more than once in the supplied list |
              enum:
                - INVALID_INPUT
                - INTERNAL_ERROR
                - INVALID_ADDRESS
                - VAULT_PAUSED
                - NOT_INITIALIZED
                - ALREADY_INITIALIZED
                - ZERO_AMOUNT
                - MATH_OVERFLOW
                - INSUFFICIENT_SHARES
                - INSUFFICIENT_UNDERLYING
                - BALANCE_MISMATCH
                - TVL_CAP_EXCEEDED
                - INVALID_WITHDRAWAL_FEE
                - FEE_EXCEEDS_MAXIMUM
                - HARVEST_COOLDOWN
                - CIRCUIT_BREAKER_TRIPPED
                - ASSET_NOT_FOUND
                - WEIGHT_MISMATCH
                - DUPLICATE_ASSET
              example: "INVALID_INPUT"
            message:
              type: string
              description: Human-readable description of the error.
              example: "positions and sources arrays are required"
            details:
              description: Optional additional context (present in development mode or for validation errors).
        meta:
          $ref: "#/components/schemas/ApiMeta"
...
  responses:
    BadRequest:
      description: Invalid or missing request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            success: false
            error:
              code: "INVALID_INPUT"
              message: "The information you provided is not valid. Please check and try again."
            meta:
              requestId: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2026-08-29T13:14:38.843Z"
              version: "1.0.0"
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          examples:
            MissingToken:
              value:
                success: false
                error:
                  code: "UNAUTHORIZED"
                  message: "Missing token"
                meta:
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
                  timestamp: "2026-08-29T13:14:38.843Z"
                  version: "1.0.0"
            InvalidToken:
              value:
                success: false
                error:
                  code: "INVALID_TOKEN"
                  message: "Invalid authentication token. Please log in again."
                meta:
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
                  timestamp: "2026-08-29T13:14:38.843Z"
                  version: "1.0.0"
            ExpiredRefresh:
              value:
                success: false
                error:
                  code: "INVALID_TOKEN"
                  message: "Invalid or expired refresh token"
                meta:
                  requestId: "550e8400-e29b-41d4-a716-446655440000"
                  timestamp: "2026-08-29T13:14:38.843Z"
                  version: "1.0.0"
    RateLimited:
      description: Too many requests — auth rate limit exceeded (20 req / 15 min)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            success: false
            error:
              code: "RATE_LIMIT_EXCEEDED"
              message: "Too many requests. Please wait a moment and try again."
            meta:
              requestId: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2026-08-29T13:14:38.843Z"
              version: "1.0.0"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "The requested resource was not found."
            meta:
              requestId: "550e8400-e29b-41d4-a716-446655440000"
              timestamp: "2026-08-29T13:14:38.843Z"
              version: "1.0.0"
...